refactor: temporarily move api provided by the json rpc to new file
